services/backend/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-emit-options --output-file=requirements/dev.txt requirements/dev.in
#
aiohttp==3.8.4
    # via web3
aiosignal==1.3.1
    # via aiohttp
aiosmtplib==2.0.1
    # via fastapi-mail
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
asttokens==2.2.1
    # via devtools
async-timeout==4.0.2
    # via aiohttp
asyncpg==0.25.0
    # via -r requirements/prod.in
attrs==22.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   pytest
base58==2.1.1
    # via multiaddr
bcrypt==4.0.0
    # via -r requirements/prod.in
bitarray==2.7.3
    # via eth-account
black==22.12.0
    # via
    #   -r requirements/dev.in
    #   fastapi-mail
blinker==1.5
    # via fastapi-mail
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   black
    #   uvicorn
coverage[toml]==7.1.0
    # via pytest-cov
cytoolz==0.12.1
    # via
    #   eth-keyfile
    #   eth-utils
devtools==0.10.0
    # via -r requirements/dev.in
dnspython==2.3.0
    # via email-validator
ecdsa==0.18.0
    # via python-jose
email-validator==1.3.1
    # via
    #   fastapi-mail
    #   pydantic
eth-abi==2.2.0
    # via
    #   eth-account
    #   web3
eth-account==0.5.9
    # via web3
eth-hash[pycryptodome]==0.3.3
    # via
    #   eth-utils
    #   web3
eth-keyfile==0.5.1
    # via eth-account
eth-keys==0.3.4
    # via
    #   eth-account
    #   eth-keyfile
eth-rlp==0.2.1
    # via
    #   eth-account
    #   web3
eth-typing==2.3.0
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==1.10.0
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
exceptiongroup==1.1.0
    # via pytest
executing==1.2.0
    # via devtools
fastapi==0.89.1
    # via -r requirements/prod.in
fastapi-mail==1.2.4
    # via -r requirements/prod.in
flake8==6.0.0
    # via -r requirements/dev.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
greenlet==2.0.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements/prod.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hexbytes==0.3.0
    # via
    #   eth-account
    #   eth-rlp
    #   web3
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via -r requirements/dev.in
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   requests
    #   rfc3986
    #   yarl
iniconfig==2.0.0
    # via pytest
ipfshttpclient==0.8.0a2
    # via web3
isort==5.12.0
    # via -r requirements/dev.in
jinja2==3.1.2
    # via fastapi-mail
jsonschema==4.17.3
    # via web3
lru-dict==1.1.8
    # via web3
markupsafe==2.1.2
    # via jinja2
mccabe==0.7.0
    # via flake8
multiaddr==0.0.9
    # via ipfshttpclient
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via black
netaddr==0.8.0
    # via multiaddr
packaging==23.0
    # via pytest
parsimonious==0.8.1
    # via eth-abi
passlib==1.7.4
    # via -r requirements/prod.in
pathspec==0.11.0
    # via black
platformdirs==2.6.2
    # via black
pluggy==1.0.0
    # via pytest
protobuf==3.19.5
    # via web3
psycopg2-binary==2.9.3
    # via -r requirements/prod.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycodestyle==2.10.0
    # via flake8
pycryptodome==3.17
    # via
    #   eth-hash
    #   eth-keyfile
pydantic[email]==1.10.4
    # via
    #   -r requirements/prod.in
    #   fastapi
    #   fastapi-mail
pyflakes==3.0.1
    # via flake8
pyrsistent==0.19.3
    # via jsonschema
pytest==7.2.1
    # via
    #   -r requirements/dev.in
    #   pytest-cov
pytest-cov==4.0.0
    # via -r requirements/dev.in
python-jose==3.3.0
    # via -r requirements/prod.in
python-multipart==0.0.5
    # via -r requirements/prod.in
requests==2.28.2
    # via
    #   ipfshttpclient
    #   web3
rfc3986[idna2008]==1.5.0
    # via httpx
rlp==2.0.1
    # via
    #   eth-account
    #   eth-rlp
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   asttokens
    #   ecdsa
    #   multiaddr
    #   parsimonious
    #   python-multipart
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.41
    # via -r requirements/prod.in
starlette==0.22.0
    # via
    #   fastapi
    #   fastapi-mail
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   pytest
toolz==0.12.0
    # via cytoolz
typing-extensions==4.4.0
    # via pydantic
urllib3==1.26.15
    # via requests
uvicorn==0.20.0
    # via -r requirements/prod.in
varint==1.0.2
    # via multiaddr
web3==5.31.3
    # via -r requirements/prod.in
websockets==9.1
    # via web3
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
